refactor of requirements complete
